Move the lxd_client package over to container/lxd.
