+++ provider/lxd/environ_raw.go
@@ -6,8 +6,8 @@ package lxd
 import (
 	"github.com/juju/errors"
 
+	"github.com/juju/juju/container/lxd/lxd_client"
 	"github.com/juju/juju/network"
-	"github.com/juju/juju/provider/lxd/lxd_client"
 )
 
 type rawProvider struct {
